Fixing N+1 query on photos index
Before this commit, @photos only contained photo objects. The view
looped over these photo objects and then queried for each of their
cities and countries, which was an N+1 query. This commit adds a scope
that includes the city and country and uses that scope in the controller
to eliminate the N+1 query.