removing random photo methods and specs
My original design for the posts page had tiles for each of the
countries that I had visited so far, padded with random photos. When I
first deployed the app, photos weren't working in production so I
commented the photo tiles out. Now photos are working, but I decided
that I like the design better without the photo tiles.