Adding counter cache to photos
Before this commit, in order to determine whether a city had any photos
to link to in the post for that city, I was using a super complicated
join query. By adding a counter cache for photos on the cities table, I
can simply include(:cities) on the initial query for the posts and then
check post.city.photos_count in the view to determine whether or not to
show the link.