changing photo-inner-container width back to 75%
Before this commit, the logo tiles were lining up with four on the top,
then three, then two. When the width is set to 75% there is a 3-3-3. It
was originally set to 76% because it was getting messed up on Windows,
but I have tested on Windows now and it's ok.